utils: utf8: change utf8_count_chars return to int
Change utf8_count_chars return type to int and drop thesys/types.h, this
way the function does not depend on posix types.

+++ lib/utils/utf8.c
@@ -8,7 +8,6 @@
 #include <string.h>
 #include <zephyr/sys/__assert.h>
 #include <errno.h>
-#include <sys/types.h>
 #include <zephyr/sys/util_utf8.h>
 
 #define ASCII_CHAR 0x7F
@@ -84,9 +83,9 @@ char *utf8_lcpy(char *dst, const char *src, size_t n)
 	return dst;
 }
 
-ssize_t utf8_count_chars(const char *s)
+int utf8_count_chars(const char *s)
 {
-	ssize_t count = 0;
+	int count = 0;
 	const char *p = s; /* getting a pointer to increment */
 
 	while (*p != '\0') {
